Input and button elements
* With button and input elements, the HTTP method is given by either the
value of the attriubte 'formmethod' or the value of the attribute
'data-enliwfen-method', where the attribute 'formmethod' takes
precedence. If neither nor of both has a non-empty value set, the
default mehtod 'GET' will be used.
* An Ajax-Call triggered by an input or a button element will add form
data if the input or button element is associated with a form and the
HTTP mehtod is set to 'POST'.