HTTP 500
Responses with HTTP code 500 may contain information about the error,
intended to be rendered into the page document. But they may as well
come with an entire HTML page, which cannot be rendered into the page
document. In this case an error dialog will be created and displayed,
where the HTML document resides in an iframe-Element.